worker/requirements.txt
requests==2.31.0
pandas==2.1.1
astor==0.8.1
pika==1.3.2
redis==5.0.1
sqlalchemy==1.4.46
psycopg2-binary==2.9.9
pydantic==1.10.13
dropbase==0.0.6
python-dotenv==1.0.0
pymysql==1.1.0
snowflake-sqlalchemy==1.5.1
